refactor: extract shared providers in a common file
